Added support for rmPP [alpha]
