+++ requirements.remote.txt
@@ -1,0 +1,3 @@
+-r requirements.txt
+paramiko==3.3.1
+psutil==5.9.5
