+++ requirements.txt
@@ -1,0 +1,3 @@
+requests==2.31.0
+loguru==0.7.2
+remarkable-update-fuse==1.0.4; sys_platform == 'linux'
